Stub initial commit.
     I'm merging content in from another, separate repository right after this
but can't create a merge commit with respect to both `main` and the branch I'm
(temporarily) using to hold a copy of that content's commit history without al-
ready having at least one initial commit on `main`.  (Well, I _could_, but
_then_ the merge commit would collapse down into a normal commit, leaving re-
sults equivlent to an unwanted fast-forward merge, as it would then only have
one parent, the one from the other/staging branch, if merged into an unborn
`main` branch.)  Add a  self-describing placeholder file to this repository to
give it content to store.  (Git repositories can't be empty.)